+++ sdk/communication/Azure.Communication.Common/api/Azure.Communication.Common.netstandard2.0.cs
@@ -36,7 +36,9 @@ namespace Azure.Communication
     }
     public partial class CommunicationTokenRefreshOptions
     {
-        public CommunicationTokenRefreshOptions(bool refreshProactively, System.Func<System.Threading.CancellationToken, string> tokenRefresher, System.Func<System.Threading.CancellationToken, System.Threading.Tasks.ValueTask<string>> asyncTokenRefresher = null, string initialToken = null) { }
+        public CommunicationTokenRefreshOptions(bool refreshProactively, System.Func<System.Threading.CancellationToken, string> tokenRefresher) { }
+        public System.Func<System.Threading.CancellationToken, System.Threading.Tasks.ValueTask<string>> AsyncTokenRefresher { get { throw null; } set { } }
+        public string InitialToken { get { throw null; } set { } }
     }
     public partial class CommunicationUserIdentifier : Azure.Communication.CommunicationIdentifier
     {